fix(ux): admin skeleton-on-error, voting_power parseFloat, touch targets
- Admin: clear skeleton containers in all 7 error catch blocks instead of
  leaving infinite shimmer animation on API failure
- Members: parseInt → parseFloat for voting_power to support fractional weights
- CSS: move WCAG 2.5.8 touch targets (min 44px) from 480px → 768px to cover
  tablets; btn-xs/btn-icon-sm get 32px minimum for dense layouts

+++ public/assets/js/pages/members.js
@@ -666,7 +666,7 @@
 
     const full_name = nameInput.value.trim();
     const email = emailInput.value.trim();
-    const voting_power = parseInt(powerInput.value) || 1;
+    const voting_power = parseFloat(powerInput.value) || 1;
     const is_active = document.getElementById('mActive').checked;
 
     const btn = document.getElementById('btnCreate');
